+++ docs/RELEASE-NOTES.md
@@ -1,13 +1,25 @@
 # Release Notes
 
+## 1.2.0 - 28 March 2022
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `KeyVaultKeyClient.SignAsync` method.
+- Added `KeyVaultKeyClient.VerifyAsync` method.
+
+Bug fixes / internal changes:
+- XML doc comment improvements.
+
 ## 1.1.0 - 31 March 2021
 
 Breaking changes:
 - (None)
 
 New features:
-- Added `KeyVaultSecretClient.SafeSetValueAsync`.
-- Added `KeyVaultSecretClient.GetSectionAsync`.
+- Added `KeyVaultSecretClient.SafeSetValueAsync` method.
+- Added `KeyVaultSecretClient.GetSectionAsync` method.
 
 Bug fixes / internal changes:
 - XML doc comment improvements.